로딩씬 이미지 추가 및 일부 스프라이트 수정

+++ ThisIsDisaster/Assets/UI/UI Scripts/StatusBarUIScript.cs
@@ -151,7 +151,7 @@ public class StatusBarUIScript : MonoBehaviour {
         Stamina = PlayerCharacter.GetComponent<CharacterModel>().CurrentStats.Stamina;
 
         HPBar.fillAmount = HP / MaxHP;
-        HPText.text = HP.ToString();
+        HPText.text = ((int)HP).ToString();
         StaminaBar.fillAmount = Stamina / MaxStamina;
         StaminaText.text = Stamina.ToString();
     }
